feat(flowbox): Inbox blir FlowBox — kön, Routing och Message log under ett namn
Magnus namn för ytan där allt som flödar in och ut hanteras: FlowBox,
i familjen FlowWink/FlowPilot/FlowWork/FlowChat. Tre flikar, ett begrepp:
Queue (kön över alla kanaler, FlowPilot först), Routing (reglerna, linser
över kanalernas egna inställningar) och Message log (liggaren). Närvaro-
toggeln i huvudet gäller alla tre. Menyn tappar tre poster.

Sidorna Inbox, Routing och Communications är borta som sidor; deras
adresser omdirigerar till rätt flik (/admin/inbox, /admin/routing,
/admin/communications → /admin/flowbox?tab=…), så gamla proveniensrader
och bokmärken landar rätt. Avsändarinställningar + brevlådor
(EmailRouterSettings) flyttar till Email → Sending, där e-postens
HOW-lager hör hemma. Vakten provider-belongs-to-the-router lär sig
FlowBox' kontrollerade flikar.
