+++ src/lib/route-prefetch.ts
@@ -37,7 +37,7 @@ const loaders: Record<string, Loader> = {
   "/admin/webinars": () => import("@/pages/admin/WebinarsPage"),
   "/admin/webmeet": () => import("@/pages/admin/WebmeetPage"),
   "/admin/forms": () => import("@/pages/admin/FormSubmissionsPage"),
-  "/admin/communications": () => import("@/pages/admin/CommunicationsPage"),
+  "/admin/flowbox": () => import("@/pages/admin/FlowBoxPage"),
   "/admin/live-support": () => import("@/pages/admin/LiveSupportPage"),
   "/admin/voice": () => import("@/pages/admin/VoicePage"),
   "/admin/tickets": () => import("@/pages/admin/TicketsPage"),
